envio de alterações no codigo: tela de avaliacao do sentimento durante tarefa + navegacao
