+++ App/components/Task.js
@@ -1,17 +1,10 @@
 import React from "react";
 import { Button, StyleSheet, Text, View } from "react-native";
-
 const Task = (props) => {
     return(
         <View style={styles.container}>
         <Text>{props.title}</Text>
         <Text>{props.text}</Text>
-
-        <View>
-            <Button title= "Não realizada" color="#FF415B" accessibilityLabel="Botão para marcar a tarefa como 'Não Realizada'"/>
-            <Button title= "Realizada" color="#00AA95"accessibilityLabel="Botão para marcar a tarefa como 'Realizada'"/>
-        </View>
-
         </View>
     )
 }
